Admin CRUD and Message implement CRUD in frontend

+++ app/Http/Controllers/Admin/MessageController.php
@@ -4,19 +4,22 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Message;
-use Illuminate\Http\Request;
 
 class MessageController extends Controller
 {
     public function index()
     {
-        $messages = Message::paginate(10); 
+        $messages = Message::paginate(10);
         return view('admin.pages.messages.index', compact('messages'));
     }
-   
-       public function destroy(Message $message)
-       {
-           $message->delete();
-           return $this->apiResponse(message: "Message Deleted Successfully");
-       }
+
+    public function destroy(Message $message)
+    {
+        $message = Message::where('id', $message->id)->first();
+        if ($message) {
+            $message->delete();
+            return to_route('messages.index')->with('success', 'Message Deleted Successfully');
+        }
+        return to_route('messages.index')->with('error', 'No Such Message');
+    }
 }
